# support a faster unique id implement