--- a/be/src/exprs/vectorized/utility_functions.cpp
+++ b/be/src/exprs/vectorized/utility_functions.cpp
@@ -2,12 +2,25 @@
 
 #include "exprs/vectorized/utility_functions.h"
 
+#include <cstdint>
+#include <cstdlib>
+#include <limits>
+#include <mutex>
+#include <random>
+#include <thread>
+
 #include "column/column_builder.h"
 #include "column/column_viewer.h"
+#include "column/vectorized_fwd.h"
+#include "common/config.h"
 #include "gen_cpp/version.h"
+#include "runtime/primitive_type.h"
 #include "runtime/runtime_state.h"
+#include "service/backend_options.h"
 #include "udf/udf_internal.h"
 #include "util/monotime.h"
+#include "util/thread.h"
+#include "util/time.h"
 #include "util/uid_util.h"
 
 namespace starrocks::vectorized {
@@ -59,6 +72,68 @@ ColumnPtr UtilityFunctions::uuid(FunctionContext*, const Columns& columns) {
     }
 
     return result.build(false);
+}
+
+inline int128_t next_uuid(int64_t timestamp, int16_t backendId, int16_t rand, int16_t tid, int16_t inc) {
+    union {
+        struct {
+            int64_t timestamp;
+            int16_t backendId;
+            int16_t rand;
+            int16_t tid;
+            int16_t inc;
+        } data;
+        int128_t res;
+    } v;
+    v.data.timestamp = timestamp;
+    v.data.backendId = backendId;
+    v.data.rand = rand;
+    v.data.tid = tid;
+    v.data.inc = inc;
+    return v.res;
+}
+
+static std::mutex uid_mutex;
+// thread ids
+// The number of executor threads is fixed.
+static std::atomic<int16_t> inc{};
+//
+static thread_local int uniq_tid = -1;
+
+int16_t get_uniq_tid() {
+    if (uniq_tid == -1) {
+        uniq_tid = inc.fetch_add(1);
+    }
+    return uniq_tid;
+}
+
+ColumnPtr UtilityFunctions::uuid_numeric(FunctionContext*, const Columns& columns) {
+    int32_t num_rows = ColumnHelper::get_const_value<TYPE_INT>(columns.back());
+    auto result = Int128Column::create(num_rows);
+
+    static std::random_device rd;
+    static std::mt19937 mt(rd());
+
+    std::uniform_int_distribution<int16_t> dist(std::numeric_limits<int16_t>::min(),
+                                                std::numeric_limits<int16_t>::max());
+
+    auto& data = result->get_data();
+
+    int backend_id = std::hash<std::string>()(BackendOptions::get_localhost());
+    backend_id ^= config::brpc_port;
+    // config::brpc_port
+    // current thread id
+    int tid = get_uniq_tid();
+    int64_t timestamp = MonotonicNanos();
+    int16_t rand = dist(mt);
+
+    DCHECK_LE(num_rows, std::numeric_limits<int16_t>::max());
+
+    for (int i = 0; i < num_rows; ++i) {
+        data[i] = next_uuid(timestamp, backend_id, rand, tid, i);
+    }
+
+    return result;
 }
 
 } // namespace starrocks::vectorized