--- a/be/src/exprs/vectorized/utility_functions.h
+++ b/be/src/exprs/vectorized/utility_functions.h
@@ -27,6 +27,7 @@ public:
      * returns uuid.
      */
     DEFINE_VECTORIZED_FN(uuid);
+    DEFINE_VECTORIZED_FN(uuid_numeric);
 };
 
 } // namespace vectorized