--- a/be/src/exprs/vectorized/utility_functions.h
+++ b/be/src/exprs/vectorized/utility_functions.h
@@ -27,6 +27,11 @@ public:
      * returns uuid.
      */
     DEFINE_VECTORIZED_FN(uuid);
+    /**
+     * Returns an approximate UUID.
+     * timestamp(64bit) + backend_id(32bit: hash(ip) ^ port) + rand (16bit) +
+     * tid(thread id 32 bit) + i (increment 16 bit)
+     */
     DEFINE_VECTORIZED_FN(uuid_numeric);
 };