## What type of PR is this：
- [ ] bug
- [x] feature
- [ ] enhancement
- [ ] others

## Which issues of this PR fixes ：
Fixes #

## Problem Summary(Required) ：

uniq_id generate rule: timestamp, backendId, rand, tid, inc

input :[]
output: TYPE_LARGEINT

mysql> select uniq_id();
+-----------------------------------+
| uniq_id()                         |
+-----------------------------------+
| 232491522003429759764239072851311 |
+-----------------------------------+
1 row in set (0.01 sec)

performace 1FE 1BE

baseline:
```
mysql> select count(*) from lineorder;
+-----------+
| count(*)  |
+-----------+
| 606039117 |
+-----------+
1 row in set (2.47 sec)
```

parallel=1
```
mysql> select count(uniq_id()) from lineorder;
+------------------+
| count(uniq_id()) |
+------------------+
|        606039117 |
+------------------+
1 row in set (4.53 sec)
```
parallel=6;

```
mysql> select count(uniq_id()) from lineorder;
+------------------+
| count(uniq_id()) |
+------------------+
|        606039117 |
+------------------+
1 row in set (2.62 sec)
```