@stdpain 2022-04-02
run starrocks_be_unittest