@stdpain 2022-04-02
run starrocks_fe_unittest