@stdpain 2022-04-06
https://github.com/Mergifyio backport branch-2.1